readOnly should affect only properties
